ENH: Important rearangement of cleanup processes
I moved the call for self.cleanup() BEFORE the call to reload
the module in the onReload function, otherwise we are cleaning
up the newly reloaded module. I also moved the call to exit
the EditorWidget to BEFORE the call to clear the mrml scene.
Exiting the EditorWidget is important to stop whatever editor
tool was in use when reloaded.