ENH: Adds functions to determine current slice number
The first step when finding the current slice was to get the RAS
from the SliceOffset. Because I am just interested in the third
RAS coordinate I was able to create a generic RAS tuple using the
sliceOffset from the labelNode. For example the RAS will be
similar to (0, 0, -439.5). The sliceOffset in this case is -439.5.
Then I could calculate the XY coordinate from the RAS using an
inverted copy of sliceNode.GetXYToRAS(). Then lastly the XY was
converted to the IJK coordinate stystem using the
labelLogic.GetXYToIJKTransform(). The current slice number is the
third value (K) in the IJK cooridinate. Once the current slice is
known, the circumference of each label in that slice can be
calculated and printed.