ENH: Renames dirDialog to two diff variables for input buttons
When there was only on self.dirDialog variable, the results and
input image list would be set to same value. Also changed how the
onSave button selects a dirName. Instead of asking the user to
select an output folder, the program creates a directory to save
the mrml scene and stats csv file based on the input results
baseDir and the current master nodes name.